Split up TestRunX
